backend/requirements.txt
fastapi 
uvicorn[standard] 
pillow 
moviepy 
pydub 
python-multipart
numpy